クラスコンポーネントから関数コンポーネントに変更とファイルの分割
