+++ test/storybook/my-app/src/components/square/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+
+interface SquarePropsInterface {
+  value: string;
+  onClick: () => void;
+}
+
+export const Square = (props:SquarePropsInterface) => {
+  return (
+    <button
+      className="square"
+      onClick={() => {
+        props.onClick();
+      }}
+    >
+      {props.value}
+    </button>
+  );
+}
